+++ argilla-server/src/argilla_server/enums.py
@@ -17,6 +17,8 @@ from enum import Enum
 
 class FieldType(str, Enum):
     text = "text"
+    image = "image"
+    chat = "chat"
 
 
 class ResponseStatus(str, Enum):