+++ argilla-server/src/argilla_server/search_engine/commons.py
@@ -148,8 +148,24 @@ def es_field_for_response_property(property: str) -> str:
 def es_mapping_for_field(field: Field) -> dict:
     field_type = field.settings["type"]
 
-    if field_type == FieldType.text:
+    if field.is_text:
         return {es_field_for_record_field(field.name): {"type": "text"}}
+    if field.is_chat:
+        es_field = {
+            "type": "object",
+            "properties": {
+                "content": {"type": "text"},
+                "role": {"type": "keyword"},
+            },
+        }
+        return {es_field_for_record_field(field.name): es_field}
+    elif field.is_image:
+        return {
+            es_field_for_record_field(field.name): {
+                "type": "object",
+                "enabled": False,
+            }
+        }
     else:
         raise Exception(f"Index configuration for field of type {field_type} cannot be generated")
 
@@ -578,6 +594,10 @@ class BaseElasticAndOpenSearchEngine(SearchEngine):
         return {
             # See https://www.elastic.co/guide/en/elasticsearch/reference/current/dynamic.html#dynamic-parameters
             "dynamic": "strict",
+            "_source": {
+                # Excluding image fields, which means they won't be even stored. See https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-source-field.html#include-exclude
+                "excludes": [es_field_for_record_field(field.name) for field in dataset.fields if field.is_image],
+            },
             "properties": {
                 # See https://www.elastic.co/guide/en/elasticsearch/reference/current/explicit-mapping.html
                 "id": {"type": "keyword"},