+++ argilla-server/CHANGELOG.md
@@ -18,7 +18,12 @@ These are the section headers that we use:
 
 ### Added
 
+- Added new `image` type dataset field supporting URLs and Data URLs. ([#5279](https://github.com/argilla-io/argilla/pull/5279))
 - Added new endpoint `GET /api/v1/datsets/:dataset_id/users/progress` to compute the users progress. ([#5367](https://github.com/argilla-io/argilla/pull/5367))
+
+### Fixed
+
+- Fixed response duplicate checking ([#5357](https://github.com/argilla-io/argilla/issues/5357))
 
 ## [2.0.0](https://github.com/argilla-io/argilla/compare/v2.0.0rc1...v2.0.0)